requirements.txt
django
python-decouple
impacket